move models from other.py into models/
