+++ models/fsrs_v5.py
@@ -1,0 +1,221 @@
+from typing import List, Union
+import torch
+from torch import nn, Tensor
+from typing import Optional
+from models.fsrs import FSRS
+
+from config import ModelConfig
+
+
+class FSRS5ParameterClipper:
+    def __init__(self, config: ModelConfig, frequency: int = 1):
+        self.frequency = frequency
+        self.config = config
+
+    def __call__(self, module):
+        if hasattr(module, "w"):
+            w = module.w.data
+            w[0] = w[0].clamp(self.config.s_min, 100)
+            w[1] = w[1].clamp(self.config.s_min, 100)
+            w[2] = w[2].clamp(self.config.s_min, 100)
+            w[3] = w[3].clamp(self.config.s_min, 100)
+            w[4] = w[4].clamp(1, 10)
+            w[5] = w[5].clamp(0.001, 4)
+            w[6] = w[6].clamp(0.001, 4)
+            w[7] = w[7].clamp(0.001, 0.75)
+            w[8] = w[8].clamp(0, 4.5)
+            w[9] = w[9].clamp(0, 0.8)
+            w[10] = w[10].clamp(0.001, 3.5)
+            w[11] = w[11].clamp(0.001, 5)
+            w[12] = w[12].clamp(0.001, 0.25)
+            w[13] = w[13].clamp(0.001, 0.9)
+            w[14] = w[14].clamp(0, 4)
+            w[15] = w[15].clamp(0, 1)
+            w[16] = w[16].clamp(1, 6)
+            w[17] = w[17].clamp(0, 2)
+            w[18] = w[18].clamp(0, 2)
+            module.w.data = w
+
+
+class FSRS5(FSRS):
+    init_w = [
+        0.40255,
+        1.18385,
+        3.173,
+        15.69105,
+        7.1949,
+        0.5345,
+        1.4604,
+        0.0046,
+        1.54575,
+        0.1192,
+        1.01925,
+        1.9395,
+        0.11,
+        0.29605,
+        2.2698,
+        0.2315,
+        2.9898,
+        0.51655,
+        0.6621,
+    ]
+    lr: float = 4e-2
+    gamma: float = 1
+    wd: float = 1e-5
+    n_epoch: int = 5
+    default_params_stddev_tensor = torch.tensor(
+        [
+            6.61,
+            9.52,
+            17.69,
+            27.74,
+            0.55,
+            0.28,
+            0.67,
+            0.12,
+            0.4,
+            0.18,
+            0.34,
+            0.27,
+            0.08,
+            0.14,
+            0.57,
+            0.25,
+            1.03,
+            0.27,
+            0.39,
+        ]
+    )
+    decay = -0.5
+    factor = 0.9 ** (1 / decay) - 1
+
+    def __init__(self, config: ModelConfig, w: Optional[List[float]] = None):
+        super(FSRS5, self).__init__(config)
+        if w is None:
+            w = self.init_w
+        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
+        self.init_w_tensor = self.w.data.clone().to(self.config.device)
+        self.clipper = FSRS5ParameterClipper(config)
+
+    def iter(
+        self,
+        sequences: Tensor,
+        delta_ts: Tensor,
+        seq_lens: Tensor,
+        real_batch_size: int,
+    ) -> dict[str, Tensor]:
+        output = super().iter(sequences, delta_ts, seq_lens, real_batch_size)
+        output["penalty"] = (
+            torch.sum(
+                torch.square(self.w - self.init_w_tensor)
+                / torch.square(self.default_params_stddev_tensor)
+            )
+            * real_batch_size
+            * self.gamma
+        )
+        return output
+
+    def forgetting_curve(self, t, s):
+        return (1 + self.factor * t / s) ** self.decay
+
+    def stability_after_success(
+        self, state: Tensor, r: Tensor, rating: Tensor
+    ) -> Tensor:
+        hard_penalty = torch.where(rating == 2, self.w[15], 1)
+        easy_bonus = torch.where(rating == 4, self.w[16], 1)
+        new_s = state[:, 0] * (
+            1
+            + torch.exp(self.w[8])
+            * (11 - state[:, 1])
+            * torch.pow(state[:, 0], -self.w[9])
+            * (torch.exp((1 - r) * self.w[10]) - 1)
+            * hard_penalty
+            * easy_bonus
+        )
+        return new_s
+
+    def stability_after_failure(self, state: Tensor, r: Tensor) -> Tensor:
+        old_s = state[:, 0]
+        new_s = (
+            self.w[11]
+            * torch.pow(state[:, 1], -self.w[12])
+            * (torch.pow(old_s + 1, self.w[13]) - 1)
+            * torch.exp((1 - r) * self.w[14])
+        )
+        new_minimum_s = old_s / torch.exp(self.w[17] * self.w[18])
+        return torch.minimum(new_s, new_minimum_s)
+
+    def stability_short_term(self, state: Tensor, rating: Tensor) -> Tensor:
+        new_s = state[:, 0] * torch.exp(self.w[17] * (rating - 3 + self.w[18]))
+        return new_s
+
+    def init_d(self, rating: Union[int, Tensor]) -> Tensor:
+        new_d = self.w[4] - torch.exp(self.w[5] * (rating - 1)) + 1
+        return new_d
+
+    def linear_damping(self, delta_d: Tensor, old_d: Tensor) -> Tensor:
+        return delta_d * (10 - old_d) / 9
+
+    def next_d(self, state: Tensor, rating: Tensor) -> Tensor:
+        delta_d = -self.w[6] * (rating - 3)
+        new_d = state[:, 1] + self.linear_damping(delta_d, state[:, 1])
+        new_d = self.mean_reversion(self.init_d(4), new_d)
+        return new_d
+
+    def step(self, X: Tensor, state: Tensor) -> Tensor:
+        """
+        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
+        :param state: shape[batch_size, 2], state[:,0] is stability, state[:,1] is difficulty
+        :return state:
+        """
+        if torch.equal(state, torch.zeros_like(state)):
+            keys = torch.tensor([1, 2, 3, 4], device=self.config.device)
+            keys = keys.view(1, -1).expand(X[:, 1].long().size(0), -1)
+            index = (X[:, 1].long().unsqueeze(1) == keys).nonzero(as_tuple=True)
+            # first learn, init memory states
+            new_s = torch.ones_like(state[:, 0], device=self.config.device)
+            new_s[index[0]] = self.w[index[1]]
+            new_d = self.init_d(X[:, 1])
+            new_d = new_d.clamp(1, 10)
+        else:
+            r = self.forgetting_curve(X[:, 0], state[:, 0])
+            short_term = X[:, 0] < 1
+            success = X[:, 1] > 1
+            new_s = torch.where(
+                short_term,
+                self.stability_short_term(state, X[:, 1]),
+                torch.where(
+                    success,
+                    self.stability_after_success(state, r, X[:, 1]),
+                    self.stability_after_failure(state, r),
+                ),
+            )
+            new_d = self.next_d(state, X[:, 1])
+            new_d = new_d.clamp(1, 10)
+        new_s = new_s.clamp(self.config.s_min, 36500)
+        return torch.stack([new_s, new_d], dim=1)
+
+    def forward(
+        self, inputs: Tensor, state: Optional[Tensor] = None
+    ) -> tuple[Tensor, Tensor]:
+        """
+        :param inputs: shape[seq_len, batch_size, 2]
+        """
+        if state is None:
+            state = torch.zeros((inputs.shape[1], 2))
+        outputs = []
+        for X in inputs:
+            state = self.step(X, state)
+            outputs.append(state)
+        return torch.stack(outputs), state
+
+    def mean_reversion(self, init: Tensor, current: Tensor) -> Tensor:
+        return self.w[7] * init + (1 - self.w[7]) * current
+
+    def state_dict(self):
+        return list(
+            map(
+                lambda x: round(float(x), 4),
+                dict(self.named_parameters())["w"].data,
+            )
+        )
