+++ other.py
@@ -4,7 +4,7 @@ import sys
 import os
 import pandas as pd
 import numpy as np
-from typing import List, Optional, Union
+from typing import Optional
 from pathlib import Path
 import matplotlib.pyplot as plt
 from concurrent.futures import ProcessPoolExecutor, as_completed
@@ -15,19 +15,38 @@ from torch import Tensor
 from sklearn.model_selection import TimeSeriesSplit  # type: ignore
 from sklearn.metrics import roc_auc_score, root_mean_squared_error, log_loss  # type: ignore
 from tqdm.auto import tqdm  # type: ignore
-from scipy.optimize import minimize  # type: ignore
 from statsmodels.nonparametric.smoothers_lowess import lowess  # type: ignore
 import warnings
 from reptile_trainer import get_inner_opt, finetune
 from script import cum_concat, remove_non_continuous_rows, remove_outliers, sort_jsonl
 import multiprocessing as mp
 import pyarrow.parquet as pq  # type: ignore
-from config import create_parser
+from config import create_parser, ModelConfig
 from utils import catch_exceptions, rmse_matrix
 
+# Import all models from /models directory
+from models.fsrs_v1 import FSRS1
+from models.fsrs_v2 import FSRS2
+from models.fsrs_v3 import FSRS3
+from models.fsrs_v4 import FSRS4
+from models.fsrs_v4dot5 import FSRS4dot5
+from models.fsrs_v5 import FSRS5
+from models.fsrs_v6 import FSRS6
+from models.rnn import RNN
+from models.gru_p import GRU_P
+from models.lstm import LSTM
+from models.transformer import Transformer
+from models.hlr import HLR
+from models.act_r import ACT_R
+from models.dash import DASH
+from models.dash_act_r import DASH_ACTR
+from models.nn_17 import NN_17
+from models.sm_2 import SM2
+from models.anki import Anki
 
 parser = create_parser()
 args, _ = parser.parse_known_args()
+model_config = ModelConfig(args)
 
 DEV_MODE = args.dev
 DRY_RUN = args.dry
@@ -125,2327 +144,6 @@ OPT_NAME = FILE_NAME + "_opt"
 S_MIN = 1e-6 if SECS_IVL else 0.01
 INIT_S_MAX = 100
 S_MAX = 36500
-
-
-class FSRS(nn.Module):
-    def __init__(self):
-        super(FSRS, self).__init__()
-
-    def forgetting_curve(self, t, s):
-        raise NotImplementedError("Forgetting curve not implemented")
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        stabilities, difficulties, *_ = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-        ].transpose(0, 1)
-        retentions = self.forgetting_curve(delta_ts, stabilities)
-        return {
-            "retentions": retentions,
-            "stabilities": stabilities,
-            "difficulties": difficulties,
-        }
-
-    def pretrain(self, train_set):
-        S0_dataset_group = (
-            train_set[train_set["i"] == 2]
-            .groupby(by=["first_rating", "delta_t"], group_keys=False)
-            .agg({"y": ["mean", "count"]})
-            .reset_index()
-        )
-        rating_stability = {}
-        rating_count = {}
-        average_recall = train_set["y"].mean()
-        r_s0_default = {str(i): self.init_w[i - 1] for i in range(1, 5)}
-
-        for first_rating in ("1", "2", "3", "4"):
-            group = S0_dataset_group[S0_dataset_group["first_rating"] == first_rating]
-            if group.empty:
-                if verbose:
-                    tqdm.write(
-                        f"Not enough data for first rating {first_rating}. Expected at least 1, got 0."
-                    )
-                continue
-            delta_t = group["delta_t"]
-            if SECS_IVL:
-                recall = group["y"]["mean"]
-            else:
-                recall = (
-                    group["y"]["mean"] * group["y"]["count"] + average_recall * 1
-                ) / (group["y"]["count"] + 1)
-            count = group["y"]["count"]
-
-            init_s0 = r_s0_default[first_rating]
-
-            def loss(stability):
-                y_pred = self.forgetting_curve(delta_t, stability)
-                logloss = sum(
-                    -(recall * np.log(y_pred) + (1 - recall) * np.log(1 - y_pred))
-                    * count
-                )
-                l1 = np.abs(stability - init_s0) / 16 if not SECS_IVL else 0
-                return logloss + l1
-
-            res = minimize(
-                loss,
-                x0=init_s0,
-                bounds=((S_MIN, INIT_S_MAX),),
-                options={"maxiter": int(sum(count))},
-            )
-            params = res.x
-            stability = params[0]
-            rating_stability[int(first_rating)] = stability
-            rating_count[int(first_rating)] = sum(count)
-
-        for small_rating, big_rating in (
-            (1, 2),
-            (2, 3),
-            (3, 4),
-            (1, 3),
-            (2, 4),
-            (1, 4),
-        ):
-            if small_rating in rating_stability and big_rating in rating_stability:
-                # if rating_count[small_rating] > 300 and rating_count[big_rating] > 300:
-                #     continue
-                if rating_stability[small_rating] > rating_stability[big_rating]:
-                    if rating_count[small_rating] > rating_count[big_rating]:
-                        rating_stability[big_rating] = rating_stability[small_rating]
-                    else:
-                        rating_stability[small_rating] = rating_stability[big_rating]
-
-        w1 = 0.41
-        w2 = 0.54
-
-        if len(rating_stability) == 0:
-            raise Exception("Not enough data for pretraining!")
-        elif len(rating_stability) == 1:
-            rating = list(rating_stability.keys())[0]
-            factor = rating_stability[rating] / r_s0_default[str(rating)]
-            init_s0 = list(map(lambda x: x * factor, r_s0_default.values()))
-        elif len(rating_stability) == 2:
-            if 1 not in rating_stability and 2 not in rating_stability:
-                rating_stability[2] = np.power(
-                    rating_stability[3], 1 / (1 - w2)
-                ) * np.power(rating_stability[4], 1 - 1 / (1 - w2))
-                rating_stability[1] = np.power(rating_stability[2], 1 / w1) * np.power(
-                    rating_stability[3], 1 - 1 / w1
-                )
-            elif 1 not in rating_stability and 3 not in rating_stability:
-                rating_stability[3] = np.power(rating_stability[2], 1 - w2) * np.power(
-                    rating_stability[4], w2
-                )
-                rating_stability[1] = np.power(rating_stability[2], 1 / w1) * np.power(
-                    rating_stability[3], 1 - 1 / w1
-                )
-            elif 1 not in rating_stability and 4 not in rating_stability:
-                rating_stability[4] = np.power(
-                    rating_stability[2], 1 - 1 / w2
-                ) * np.power(rating_stability[3], 1 / w2)
-                rating_stability[1] = np.power(rating_stability[2], 1 / w1) * np.power(
-                    rating_stability[3], 1 - 1 / w1
-                )
-            elif 2 not in rating_stability and 3 not in rating_stability:
-                rating_stability[2] = np.power(
-                    rating_stability[1], w1 / (w1 + w2 - w1 * w2)
-                ) * np.power(rating_stability[4], 1 - w1 / (w1 + w2 - w1 * w2))
-                rating_stability[3] = np.power(
-                    rating_stability[1], 1 - w2 / (w1 + w2 - w1 * w2)
-                ) * np.power(rating_stability[4], w2 / (w1 + w2 - w1 * w2))
-            elif 2 not in rating_stability and 4 not in rating_stability:
-                rating_stability[2] = np.power(rating_stability[1], w1) * np.power(
-                    rating_stability[3], 1 - w1
-                )
-                rating_stability[4] = np.power(
-                    rating_stability[2], 1 - 1 / w2
-                ) * np.power(rating_stability[3], 1 / w2)
-            elif 3 not in rating_stability and 4 not in rating_stability:
-                rating_stability[3] = np.power(
-                    rating_stability[1], 1 - 1 / (1 - w1)
-                ) * np.power(rating_stability[2], 1 / (1 - w1))
-                rating_stability[4] = np.power(
-                    rating_stability[2], 1 - 1 / w2
-                ) * np.power(rating_stability[3], 1 / w2)
-            init_s0 = [
-                item[1] for item in sorted(rating_stability.items(), key=lambda x: x[0])
-            ]
-        elif len(rating_stability) == 3:
-            if 1 not in rating_stability:
-                rating_stability[1] = np.power(rating_stability[2], 1 / w1) * np.power(
-                    rating_stability[3], 1 - 1 / w1
-                )
-            elif 2 not in rating_stability:
-                rating_stability[2] = np.power(rating_stability[1], w1) * np.power(
-                    rating_stability[3], 1 - w1
-                )
-            elif 3 not in rating_stability:
-                rating_stability[3] = np.power(rating_stability[2], 1 - w2) * np.power(
-                    rating_stability[4], w2
-                )
-            elif 4 not in rating_stability:
-                rating_stability[4] = np.power(
-                    rating_stability[2], 1 - 1 / w2
-                ) * np.power(rating_stability[3], 1 / w2)
-            init_s0 = [
-                item[1] for item in sorted(rating_stability.items(), key=lambda x: x[0])
-            ]
-        elif len(rating_stability) == 4:
-            init_s0 = [
-                item[1] for item in sorted(rating_stability.items(), key=lambda x: x[0])
-            ]
-        self.w.data[0:4] = Tensor(
-            list(map(lambda x: max(min(INIT_S_MAX, x), S_MIN), init_s0))
-        )
-        self.init_w_tensor = self.w.data.clone().to(DEVICE)
-
-
-class FSRS1ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp(0.1, 10)
-            w[1] = w[1].clamp(1, 10)
-            w[2] = w[2].clamp(0.01, 10)
-            w[3] = w[3].clamp(-1, -0.01)
-            w[4] = w[4].clamp(-1, -0.01)
-            w[5] = w[5].clamp(0.01, 10)
-            w[6] = w[6].clamp(-1, -0.01)
-            module.w.data = w
-
-
-class FSRS1(FSRS):
-    # 7 params
-    init_w = [2, 5, 3, -0.7, -0.2, 1, -0.3]
-    clipper = FSRS1ParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(FSRS1, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def forgetting_curve(self, t, s):
-        return 0.9 ** (t / s)
-
-    def stability_after_success(
-        self, state: Tensor, new_d: Tensor, r: Tensor
-    ) -> Tensor:
-        new_s = state[:, 0] * (
-            1
-            + torch.exp(self.w[2])
-            * torch.pow(new_d + 0.1, self.w[3])
-            * torch.pow(state[:, 0], self.w[4])
-            * (torch.exp((1 - r) * self.w[5]) - 1)
-        )
-        return new_s
-
-    def stability_after_failure(self, lapses: Tensor) -> Tensor:
-        new_s = self.w[0] * torch.exp(self.w[6] * lapses)
-        return new_s
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 3], state[:,0] is stability, state[:,1] is difficulty, state[:,2] is the number of lapses
-        :return state:
-        """
-        if torch.equal(state, torch.zeros_like(state)):
-            # first learn, init memory states
-            new_s = self.w[0] * 0.25 * torch.pow(2, X[:, 1] - 1)
-            new_d = self.w[1] - X[:, 1] + 3
-            new_l = torch.relu(2 - X[:, 1])
-        else:
-            r = self.forgetting_curve(X[:, 0], state[:, 0])
-            new_d = torch.relu(state[:, 1] + r - 0.25 * torch.pow(2, X[:, 1] - 1) + 0.1)
-            condition = X[:, 1] > 1
-            new_s = torch.where(
-                condition,
-                self.stability_after_success(state, new_d, r),
-                self.stability_after_failure(state[:, 2]),
-            )
-            new_l = state[:, 2] + torch.relu(2 - X[:, 1])
-        new_s = new_s.clamp(S_MIN, S_MAX)
-        return torch.stack([new_s, new_d, new_l], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 3))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class FSRS2ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp(0.1, 10)
-            w[1] = w[1].clamp(0.01, 10)
-            w[2] = w[2].clamp(1, 10)
-            w[3] = w[3].clamp(-10, -0.01)
-            w[4] = w[4].clamp(-10, -0.01)
-            w[5] = w[5].clamp(0, 1)
-            w[6] = w[6].clamp(0, 5)
-            w[7] = w[7].clamp(-2, -0.01)
-            w[8] = w[8].clamp(-2, -0.01)
-            w[9] = w[9].clamp(0.01, 2)
-            w[10] = w[10].clamp(0, 5)
-            w[11] = w[11].clamp(-2, -0.01)
-            w[12] = w[12].clamp(0.01, 1)
-            w[13] = w[13].clamp(0.01, 2)
-            module.w.data = w
-
-
-class FSRS2(FSRS):
-    # 14 params
-    init_w = [1, 1, 1, -1, -1, 0.2, 3, -0.8, -0.2, 1.3, 2.6, -0.2, 0.6, 1.5]
-    clipper = FSRS2ParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(FSRS2, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def forgetting_curve(self, t, s):
-        return 0.9 ** (t / s)
-
-    def stability_after_success(
-        self, state: Tensor, new_d: Tensor, r: Tensor
-    ) -> Tensor:
-        new_s = state[:, 0] * (
-            1
-            + torch.exp(self.w[6])
-            * torch.pow(new_d, self.w[7])
-            * torch.pow(state[:, 0], self.w[8])
-            * (torch.exp((1 - r) * self.w[9]) - 1)
-        )
-        return new_s
-
-    def stability_after_failure(
-        self, state: Tensor, new_d: Tensor, r: Tensor
-    ) -> Tensor:
-        new_s = (
-            self.w[10]
-            * torch.pow(new_d, self.w[11])
-            * torch.pow(state[:, 0], self.w[12])
-            * (torch.exp((1 - r) * self.w[13]) - 1)
-        )
-        return new_s
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 2], state[:,0] is stability, state[:,1] is difficulty
-        :return state:
-        """
-        if torch.equal(state, torch.zeros_like(state)):
-            # first learn, init memory states
-            new_s = self.w[0] * (self.w[1] * (X[:, 1] - 1) + 1)
-            new_d = self.w[2] * (self.w[3] * (X[:, 1] - 4) + 1)
-            new_d = new_d.clamp(1, 10)
-        else:
-            r = self.forgetting_curve(X[:, 0], state[:, 0])
-            new_d = state[:, 1] + self.w[4] * (X[:, 1] - 3)
-            new_d = self.mean_reversion(self.w[2] * (-self.w[3] + 1), new_d)
-            new_d = new_d.clamp(1, 10)
-            condition = X[:, 1] > 1
-            new_s = torch.where(
-                condition,
-                self.stability_after_success(state, new_d, r),
-                self.stability_after_failure(state, new_d, r),
-            )
-        new_s = new_s.clamp(S_MIN, S_MAX)
-        return torch.stack([new_s, new_d], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 2))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def mean_reversion(self, init: Tensor, current: Tensor) -> Tensor:
-        return self.w[5] * init + (1 - self.w[5]) * current
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class FSRS3ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp(0.1, 10)
-            w[1] = w[1].clamp(0.1, 5)
-            w[2] = w[2].clamp(1, 10)
-            w[3] = w[3].clamp(-5, -0.1)
-            w[4] = w[4].clamp(-5, -0.1)
-            w[5] = w[5].clamp(0.05, 0.5)
-            w[6] = w[6].clamp(0, 2)
-            w[7] = w[7].clamp(-0.8, -0.15)
-            w[8] = w[8].clamp(0.01, 1.5)
-            w[9] = w[9].clamp(0.5, 5)
-            w[10] = w[10].clamp(-2, -0.01)
-            w[11] = w[11].clamp(0.01, 0.9)
-            w[12] = w[12].clamp(0.01, 2)
-            module.w.data = w
-
-
-class FSRS3(FSRS):
-    # 13 params
-    init_w = [
-        0.9605,
-        1.7234,
-        4.8527,
-        -1.1917,
-        -1.2956,
-        0.0573,
-        1.7352,
-        -0.1673,
-        1.065,
-        1.8907,
-        -0.3832,
-        0.5867,
-        1.0721,
-    ]
-    clipper = FSRS3ParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(FSRS3, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def forgetting_curve(self, t, s):
-        return 0.9 ** (t / s)
-
-    def stability_after_success(
-        self, state: Tensor, new_d: Tensor, r: Tensor
-    ) -> Tensor:
-        new_s = state[:, 0] * (
-            1
-            + torch.exp(self.w[6])
-            * (11 - new_d)
-            * torch.pow(state[:, 0], self.w[7])
-            * (torch.exp((1 - r) * self.w[8]) - 1)
-        )
-        return new_s
-
-    def stability_after_failure(
-        self, state: Tensor, new_d: Tensor, r: Tensor
-    ) -> Tensor:
-        new_s = (
-            self.w[9]
-            * torch.pow(new_d, self.w[10])
-            * torch.pow(state[:, 0], self.w[11])
-            * torch.exp((1 - r) * self.w[12])
-        )
-        return new_s
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 2], state[:,0] is stability, state[:,1] is difficulty
-        :return state:
-        """
-        if torch.equal(state, torch.zeros_like(state)):
-            # first learn, init memory states
-            new_s = self.w[0] + self.w[1] * (X[:, 1] - 1)
-            new_d = self.w[2] + self.w[3] * (X[:, 1] - 3)
-            new_d = new_d.clamp(1, 10)
-        else:
-            r = self.forgetting_curve(X[:, 0], state[:, 0])
-            new_d = state[:, 1] + self.w[4] * (X[:, 1] - 3)
-            new_d = self.mean_reversion(self.w[2], new_d)
-            new_d = new_d.clamp(1, 10)
-            condition = X[:, 1] > 1
-            new_s = torch.where(
-                condition,
-                self.stability_after_success(state, new_d, r),
-                self.stability_after_failure(state, new_d, r),
-            )
-        new_s = new_s.clamp(S_MIN, S_MAX)
-        return torch.stack([new_s, new_d], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 2))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def mean_reversion(self, init: Tensor, current: Tensor) -> Tensor:
-        return self.w[5] * init + (1 - self.w[5]) * current
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class FSRS4ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[4] = w[4].clamp(1, 10)
-            w[5] = w[5].clamp(0.1, 5)
-            w[6] = w[6].clamp(0.1, 5)
-            w[7] = w[7].clamp(0, 0.5)
-            w[8] = w[8].clamp(0, 3)
-            w[9] = w[9].clamp(0.1, 0.8)
-            w[10] = w[10].clamp(0.01, 2.5)
-            w[11] = w[11].clamp(0.5, 5)
-            w[12] = w[12].clamp(0.01, 0.2)
-            w[13] = w[13].clamp(0.01, 0.9)
-            w[14] = w[14].clamp(0.01, 2)
-            w[15] = w[15].clamp(0, 1)
-            w[16] = w[16].clamp(1, 4)
-            module.w.data = w
-
-
-class FSRS4(FSRS):
-    # 17 params
-    init_w = [
-        0.4,
-        0.9,
-        2.3,
-        10.9,
-        4.93,
-        0.94,
-        0.86,
-        0.01,
-        1.49,
-        0.14,
-        0.94,
-        2.18,
-        0.05,
-        0.34,
-        1.26,
-        0.29,
-        2.61,
-    ]
-    clipper = FSRS4ParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(FSRS4, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def forgetting_curve(self, t, s):
-        return (1 + t / (9 * s)) ** -1
-
-    def stability_after_success(
-        self, state: Tensor, r: Tensor, rating: Tensor
-    ) -> Tensor:
-        hard_penalty = torch.where(rating == 2, self.w[15], 1)
-        easy_bonus = torch.where(rating == 4, self.w[16], 1)
-        new_s = state[:, 0] * (
-            1
-            + torch.exp(self.w[8])
-            * (11 - state[:, 1])
-            * torch.pow(state[:, 0], -self.w[9])
-            * (torch.exp((1 - r) * self.w[10]) - 1)
-            * hard_penalty
-            * easy_bonus
-        )
-        return new_s
-
-    def stability_after_failure(self, state: Tensor, r: Tensor) -> Tensor:
-        new_s = (
-            self.w[11]
-            * torch.pow(state[:, 1], -self.w[12])
-            * (torch.pow(state[:, 0] + 1, self.w[13]) - 1)
-            * torch.exp((1 - r) * self.w[14])
-        )
-        return new_s
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 2], state[:,0] is stability, state[:,1] is difficulty
-        :return state:
-        """
-        if torch.equal(state, torch.zeros_like(state)):
-            keys = torch.tensor([1, 2, 3, 4], device=DEVICE)
-            keys = keys.view(1, -1).expand(X[:, 1].long().size(0), -1)
-            index = (X[:, 1].long().unsqueeze(1) == keys).nonzero(as_tuple=True)
-            # first learn, init memory states
-            new_s = torch.ones_like(state[:, 0], device=DEVICE)
-            new_s[index[0]] = self.w[index[1]]
-            new_d = self.w[4] - self.w[5] * (X[:, 1] - 3)
-            new_d = new_d.clamp(1, 10)
-        else:
-            r = self.forgetting_curve(X[:, 0], state[:, 0])
-            condition = X[:, 1] > 1
-            new_s = torch.where(
-                condition,
-                self.stability_after_success(state, r, X[:, 1]),
-                self.stability_after_failure(state, r),
-            )
-            new_d = state[:, 1] - self.w[6] * (X[:, 1] - 3)
-            new_d = self.mean_reversion(self.w[4], new_d)
-            new_d = new_d.clamp(1, 10)
-        new_s = new_s.clamp(S_MIN, S_MAX)
-        return torch.stack([new_s, new_d], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 2))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def mean_reversion(self, init: Tensor, current: Tensor) -> Tensor:
-        return self.w[7] * init + (1 - self.w[7]) * current
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class FSRS4dot5ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[4] = w[4].clamp(0, 10)
-            w[5] = w[5].clamp(0.01, 5)
-            w[6] = w[6].clamp(0.01, 5)
-            w[7] = w[7].clamp(0, 0.8)
-            w[8] = w[8].clamp(0, 6)
-            w[9] = w[9].clamp(0, 0.8)
-            w[10] = w[10].clamp(0.01, 5)
-            w[11] = w[11].clamp(0.2, 6)
-            w[12] = w[12].clamp(0.01, 0.4)
-            w[13] = w[13].clamp(0.01, 0.9)
-            w[14] = w[14].clamp(0.01, 4)
-            w[15] = w[15].clamp(0, 1)
-            w[16] = w[16].clamp(1, 10)
-            module.w.data = w
-
-
-DECAY = -0.5
-FACTOR = 0.9 ** (1 / DECAY) - 1
-
-
-class FSRS4dot5(FSRS):
-    # 17 params
-    init_w = (
-        [
-            0.4872,
-            1.4003,
-            3.7145,
-            13.8206,
-            5.1618,
-            1.2298,
-            0.8975,
-            0.031,
-            1.6474,
-            0.1367,
-            1.0461,
-            2.1072,
-            0.0793,
-            0.3246,
-            1.587,
-            0.2272,
-            2.8755,
-        ]
-        if not SECS_IVL
-        else [
-            0.0012,
-            0.0826,
-            0.8382,
-            26.2146,
-            4.8622,
-            1.0311,
-            0.8295,
-            0.0379,
-            2.0884,
-            0.4704,
-            1.2009,
-            1.7196,
-            0.1874,
-            0.1593,
-            1.5636,
-            0.2358,
-            3.3175,
-        ]
-    )
-    clipper = FSRS4dot5ParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(FSRS4dot5, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def stability_after_success(
-        self, state: Tensor, r: Tensor, rating: Tensor
-    ) -> Tensor:
-        hard_penalty = torch.where(rating == 2, self.w[15], 1)
-        easy_bonus = torch.where(rating == 4, self.w[16], 1)
-        new_s = state[:, 0] * (
-            1
-            + torch.exp(self.w[8])
-            * (11 - state[:, 1])
-            * torch.pow(state[:, 0], -self.w[9])
-            * (torch.exp((1 - r) * self.w[10]) - 1)
-            * hard_penalty
-            * easy_bonus
-        )
-        return new_s
-
-    def stability_after_failure(self, state: Tensor, r: Tensor) -> Tensor:
-        new_s = (
-            self.w[11]
-            * torch.pow(state[:, 1], -self.w[12])
-            * (torch.pow(state[:, 0] + 1, self.w[13]) - 1)
-            * torch.exp((1 - r) * self.w[14])
-        )
-        return torch.minimum(new_s, state[:, 0])
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 2], state[:,0] is stability, state[:,1] is difficulty
-        :return state:
-        """
-        if torch.equal(state, torch.zeros_like(state)):
-            keys = torch.tensor([1, 2, 3, 4], device=DEVICE)
-            keys = keys.view(1, -1).expand(X[:, 1].long().size(0), -1)
-            index = (X[:, 1].long().unsqueeze(1) == keys).nonzero(as_tuple=True)
-            # first learn, init memory states
-            new_s = torch.ones_like(state[:, 0], device=DEVICE)
-            new_s[index[0]] = self.w[index[1]]
-            new_d = self.w[4] - self.w[5] * (X[:, 1] - 3)
-            new_d = new_d.clamp(1, 10)
-        else:
-            r = self.forgetting_curve(X[:, 0], state[:, 0])
-            condition = X[:, 1] > 1
-            new_s = torch.where(
-                condition,
-                self.stability_after_success(state, r, X[:, 1]),
-                self.stability_after_failure(state, r),
-            )
-            new_d = state[:, 1] - self.w[6] * (X[:, 1] - 3)
-            new_d = self.mean_reversion(self.w[4], new_d)
-            new_d = new_d.clamp(1, 10)
-        new_s = new_s.clamp(S_MIN, S_MAX)
-        return torch.stack([new_s, new_d], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 2))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def mean_reversion(self, init: Tensor, current: Tensor) -> Tensor:
-        return self.w[7] * init + (1 - self.w[7]) * current
-
-    def forgetting_curve(self, t, s):
-        return (1 + FACTOR * t / s) ** DECAY
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 6),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class FSRS5ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp(S_MIN, 100)
-            w[1] = w[1].clamp(S_MIN, 100)
-            w[2] = w[2].clamp(S_MIN, 100)
-            w[3] = w[3].clamp(S_MIN, 100)
-            w[4] = w[4].clamp(1, 10)
-            w[5] = w[5].clamp(0.001, 4)
-            w[6] = w[6].clamp(0.001, 4)
-            w[7] = w[7].clamp(0.001, 0.75)
-            w[8] = w[8].clamp(0, 4.5)
-            w[9] = w[9].clamp(0, 0.8)
-            w[10] = w[10].clamp(0.001, 3.5)
-            w[11] = w[11].clamp(0.001, 5)
-            w[12] = w[12].clamp(0.001, 0.25)
-            w[13] = w[13].clamp(0.001, 0.9)
-            w[14] = w[14].clamp(0, 4)
-            w[15] = w[15].clamp(0, 1)
-            w[16] = w[16].clamp(1, 6)
-            w[17] = w[17].clamp(0, 2)
-            w[18] = w[18].clamp(0, 2)
-            module.w.data = w
-
-
-class FSRS5(FSRS):
-    init_w = [
-        0.40255,
-        1.18385,
-        3.173,
-        15.69105,
-        7.1949,
-        0.5345,
-        1.4604,
-        0.0046,
-        1.54575,
-        0.1192,
-        1.01925,
-        1.9395,
-        0.11,
-        0.29605,
-        2.2698,
-        0.2315,
-        2.9898,
-        0.51655,
-        0.6621,
-    ]
-    clipper = FSRS5ParameterClipper()
-    lr: float = 4e-2
-    gamma: float = 1
-    wd: float = 1e-5
-    n_epoch: int = 5
-    default_params_stddev_tensor = torch.tensor(
-        [
-            6.61,
-            9.52,
-            17.69,
-            27.74,
-            0.55,
-            0.28,
-            0.67,
-            0.12,
-            0.4,
-            0.18,
-            0.34,
-            0.27,
-            0.08,
-            0.14,
-            0.57,
-            0.25,
-            1.03,
-            0.27,
-            0.39,
-        ]
-    )
-
-    def __init__(self, w: List[float] = init_w):
-        super(FSRS5, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-        self.init_w_tensor = self.w.data.clone().to(DEVICE)
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        output = super().iter(sequences, delta_ts, seq_lens, real_batch_size)
-        output["penalty"] = (
-            torch.sum(
-                torch.square(self.w - self.init_w_tensor)
-                / torch.square(self.default_params_stddev_tensor)
-            )
-            * real_batch_size
-            * self.gamma
-        )
-        return output
-
-    def forgetting_curve(self, t, s):
-        return (1 + FACTOR * t / s) ** DECAY
-
-    def stability_after_success(
-        self, state: Tensor, r: Tensor, rating: Tensor
-    ) -> Tensor:
-        hard_penalty = torch.where(rating == 2, self.w[15], 1)
-        easy_bonus = torch.where(rating == 4, self.w[16], 1)
-        new_s = state[:, 0] * (
-            1
-            + torch.exp(self.w[8])
-            * (11 - state[:, 1])
-            * torch.pow(state[:, 0], -self.w[9])
-            * (torch.exp((1 - r) * self.w[10]) - 1)
-            * hard_penalty
-            * easy_bonus
-        )
-        return new_s
-
-    def stability_after_failure(self, state: Tensor, r: Tensor) -> Tensor:
-        old_s = state[:, 0]
-        new_s = (
-            self.w[11]
-            * torch.pow(state[:, 1], -self.w[12])
-            * (torch.pow(old_s + 1, self.w[13]) - 1)
-            * torch.exp((1 - r) * self.w[14])
-        )
-        new_minimum_s = old_s / torch.exp(self.w[17] * self.w[18])
-        return torch.minimum(new_s, new_minimum_s)
-
-    def stability_short_term(self, state: Tensor, rating: Tensor) -> Tensor:
-        new_s = state[:, 0] * torch.exp(self.w[17] * (rating - 3 + self.w[18]))
-        return new_s
-
-    def init_d(self, rating: Union[int, Tensor]) -> Tensor:
-        new_d = self.w[4] - torch.exp(self.w[5] * (rating - 1)) + 1
-        return new_d
-
-    def linear_damping(self, delta_d: Tensor, old_d: Tensor) -> Tensor:
-        return delta_d * (10 - old_d) / 9
-
-    def next_d(self, state: Tensor, rating: Tensor) -> Tensor:
-        delta_d = -self.w[6] * (rating - 3)
-        new_d = state[:, 1] + self.linear_damping(delta_d, state[:, 1])
-        new_d = self.mean_reversion(self.init_d(4), new_d)
-        return new_d
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 2], state[:,0] is stability, state[:,1] is difficulty
-        :return state:
-        """
-        if torch.equal(state, torch.zeros_like(state)):
-            keys = torch.tensor([1, 2, 3, 4], device=DEVICE)
-            keys = keys.view(1, -1).expand(X[:, 1].long().size(0), -1)
-            index = (X[:, 1].long().unsqueeze(1) == keys).nonzero(as_tuple=True)
-            # first learn, init memory states
-            new_s = torch.ones_like(state[:, 0], device=DEVICE)
-            new_s[index[0]] = self.w[index[1]]
-            new_d = self.init_d(X[:, 1])
-            new_d = new_d.clamp(1, 10)
-        else:
-            r = self.forgetting_curve(X[:, 0], state[:, 0])
-            short_term = X[:, 0] < 1
-            success = X[:, 1] > 1
-            new_s = torch.where(
-                short_term,
-                self.stability_short_term(state, X[:, 1]),
-                torch.where(
-                    success,
-                    self.stability_after_success(state, r, X[:, 1]),
-                    self.stability_after_failure(state, r),
-                ),
-            )
-            new_d = self.next_d(state, X[:, 1])
-            new_d = new_d.clamp(1, 10)
-        new_s = new_s.clamp(S_MIN, 36500)
-        return torch.stack([new_s, new_d], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 2))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def mean_reversion(self, init: Tensor, current: Tensor) -> Tensor:
-        return self.w[7] * init + (1 - self.w[7]) * current
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class FSRS6ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp(S_MIN, 100)
-            w[1] = w[1].clamp(S_MIN, 100)
-            w[2] = w[2].clamp(S_MIN, 100)
-            w[3] = w[3].clamp(S_MIN, 100)
-            w[4] = w[4].clamp(1, 10)
-            w[5] = w[5].clamp(0.001, 4)
-            w[6] = w[6].clamp(0.001, 4)
-            w[7] = w[7].clamp(0.001, 0.75)
-            w[8] = w[8].clamp(0, 4.5)
-            w[9] = w[9].clamp(0, 0.8)
-            w[10] = w[10].clamp(0.001, 3.5)
-            w[11] = w[11].clamp(0.001, 5)
-            w[12] = w[12].clamp(0.001, 0.25)
-            w[13] = w[13].clamp(0.001, 0.9)
-            w[14] = w[14].clamp(0, 4)
-            w[15] = w[15].clamp(0, 1)
-            w[16] = w[16].clamp(1, 6)
-            w[17] = w[17].clamp(0, 2)
-            w[18] = w[18].clamp(0, 2)
-            w[19] = w[19].clamp(0, 0.8)
-            w[20] = w[20].clamp(0.1, 0.8)
-            module.w.data = w
-
-
-class FSRS6(FSRS):
-    init_w = [
-        0.212,
-        1.2931,
-        2.3065,
-        8.2956,
-        6.4133,
-        0.8334,
-        3.0194,
-        0.001,
-        1.8722,
-        0.1666,
-        0.796,
-        1.4835,
-        0.0614,
-        0.2629,
-        1.6483,
-        0.6014,
-        1.8729,
-        0.5425,
-        0.0912,
-        0.0658,
-        0.1542,
-    ]
-    clipper = FSRS6ParameterClipper()
-    lr: float = 4e-2
-    gamma: float = 1
-    wd: float = 1e-5
-    n_epoch: int = 5
-    default_params_stddev_tensor = torch.tensor(
-        [
-            6.43,
-            9.66,
-            17.58,
-            27.85,
-            0.57,
-            0.28,
-            0.6,
-            0.12,
-            0.39,
-            0.18,
-            0.33,
-            0.3,
-            0.09,
-            0.16,
-            0.57,
-            0.25,
-            1.03,
-            0.31,
-            0.32,
-            0.14,
-            0.27,
-        ]
-    )
-
-    def __init__(self, w: List[float] = init_w):
-        super(FSRS6, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-        self.init_w_tensor = self.w.data.clone().to(DEVICE)
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        stabilities, difficulties = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-        ].transpose(0, 1)
-        retentions = self.forgetting_curve(delta_ts, stabilities, -self.w[20])
-        output = {
-            "retentions": retentions,
-            "stabilities": stabilities,
-            "difficulties": difficulties,
-        }
-        output["penalty"] = (
-            torch.sum(
-                torch.square(self.w - self.init_w_tensor)
-                / torch.square(self.default_params_stddev_tensor)
-            )
-            * real_batch_size
-            * self.gamma
-        )
-        return output
-
-    def forgetting_curve(self, t, s, decay=-0.2):
-        factor = 0.9 ** (1 / decay) - 1
-        return (1 + factor * t / s) ** decay
-
-    def stability_after_success(
-        self, state: Tensor, r: Tensor, rating: Tensor
-    ) -> Tensor:
-        hard_penalty = torch.where(rating == 2, self.w[15], 1)
-        easy_bonus = torch.where(rating == 4, self.w[16], 1)
-        new_s = state[:, 0] * (
-            1
-            + torch.exp(self.w[8])
-            * (11 - state[:, 1])
-            * torch.pow(state[:, 0], -self.w[9])
-            * (torch.exp((1 - r) * self.w[10]) - 1)
-            * hard_penalty
-            * easy_bonus
-        )
-        return new_s
-
-    def stability_after_failure(self, state: Tensor, r: Tensor) -> Tensor:
-        old_s = state[:, 0]
-        new_s = (
-            self.w[11]
-            * torch.pow(state[:, 1], -self.w[12])
-            * (torch.pow(old_s + 1, self.w[13]) - 1)
-            * torch.exp((1 - r) * self.w[14])
-        )
-        new_minimum_s = old_s / torch.exp(self.w[17] * self.w[18])
-        return torch.minimum(new_s, new_minimum_s)
-
-    def stability_short_term(self, state: Tensor, rating: Tensor) -> Tensor:
-        sinc = torch.exp(self.w[17] * (rating - 3 + self.w[18])) * torch.pow(
-            state[:, 0], -self.w[19]
-        )
-        new_s = state[:, 0] * torch.where(rating >= 3, sinc.clamp(min=1), sinc)
-        return new_s
-
-    def init_d(self, rating: Union[int, Tensor]) -> Tensor:
-        new_d = self.w[4] - torch.exp(self.w[5] * (rating - 1)) + 1
-        return new_d
-
-    def linear_damping(self, delta_d: Tensor, old_d: Tensor) -> Tensor:
-        return delta_d * (10 - old_d) / 9
-
-    def next_d(self, state: Tensor, rating: Tensor) -> Tensor:
-        delta_d = -self.w[6] * (rating - 3)
-        new_d = state[:, 1] + self.linear_damping(delta_d, state[:, 1])
-        new_d = self.mean_reversion(self.init_d(4), new_d)
-        return new_d
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 2], state[:,0] is stability, state[:,1] is difficulty
-        :return state:
-        """
-        if torch.equal(state, torch.zeros_like(state)):
-            keys = torch.tensor([1, 2, 3, 4], device=DEVICE)
-            keys = keys.view(1, -1).expand(X[:, 1].long().size(0), -1)
-            index = (X[:, 1].long().unsqueeze(1) == keys).nonzero(as_tuple=True)
-            # first learn, init memory states
-            new_s = torch.ones_like(state[:, 0], device=DEVICE)
-            new_s[index[0]] = self.w[index[1]]
-            new_d = self.init_d(X[:, 1])
-            new_d = new_d.clamp(1, 10)
-        else:
-            r = self.forgetting_curve(X[:, 0], state[:, 0], -self.w[20])
-            short_term = X[:, 0] < 1
-            success = X[:, 1] > 1
-            new_s = torch.where(
-                short_term,
-                self.stability_short_term(state, X[:, 1]),
-                torch.where(
-                    success,
-                    self.stability_after_success(state, r, X[:, 1]),
-                    self.stability_after_failure(state, r),
-                ),
-            )
-            new_d = self.next_d(state, X[:, 1])
-            new_d = new_d.clamp(1, 10)
-        new_s = new_s.clamp(S_MIN, 36500)
-        return torch.stack([new_s, new_d], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 2))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def mean_reversion(self, init: Tensor, current: Tensor) -> Tensor:
-        return self.w[7] * init + (1 - self.w[7]) * current
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class RNN(nn.Module):
-    # 39 params with default settings
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, state_dict=None):
-        super().__init__()
-        self.n_input = 2
-        self.n_hidden = 2
-        self.n_out = 1
-        self.n_layers = 1
-        if MODEL_NAME == "GRU":
-            self.rnn = nn.GRU(
-                input_size=self.n_input,
-                hidden_size=self.n_hidden,
-                num_layers=self.n_layers,
-            )
-            nn.init.orthogonal_(self.rnn.weight_ih_l0)
-            nn.init.orthogonal_(self.rnn.weight_hh_l0)
-            self.rnn.bias_ih_l0.data.fill_(0)
-            self.rnn.bias_hh_l0.data.fill_(0)
-        else:
-            self.rnn = nn.RNN(
-                input_size=self.n_input,
-                hidden_size=self.n_hidden,
-                num_layers=self.n_layers,
-            )
-
-        self.fc = nn.Linear(self.n_hidden, self.n_out)
-
-        if state_dict is not None:
-            self.load_state_dict(state_dict)
-        else:
-            try:
-                self.load_state_dict(
-                    torch.load(
-                        f"./pretrain/{FILE_NAME}_pretrain.pth",
-                        weights_only=True,
-                        map_location=DEVICE,
-                    )
-                )
-            except FileNotFoundError:
-                pass
-
-    def forward(self, x, hx=None):
-        x, h = self.rnn(x, hx=hx)
-        output = torch.exp(self.fc(x))
-        return output, h
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        stabilities = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-            0,
-        ]
-        retentions = self.forgetting_curve(delta_ts, stabilities)
-        return {"retentions": retentions, "stabilities": stabilities}
-
-    def full_connect(self, h):
-        return self.fc(h)
-
-    def forgetting_curve(self, t, s):
-        return (1 + FACTOR * t / s) ** DECAY
-
-
-class ResBlock(torch.nn.Module):
-    def __init__(self, module):
-        super().__init__()
-        self.module = module
-
-    def forward(self, inputs):
-        return self.module(inputs) + inputs
-
-
-class RNNWrapper(torch.nn.Module):
-    def __init__(self, module):
-        super().__init__()
-        self.module = module
-
-    def forward(self, inputs):
-        outputs, _ = self.module(inputs)
-        return outputs
-
-
-class LSTM(nn.Module):
-    """
-    This model is trained with reptile_trainer.py, and was run with the flags
-    ['--short', '--secs', '--equalize_test_with_non_secs' '--processes 2']
-    It uses:
-    - same-day reviews as features
-    - fractional intervals
-    - the duration of each review as an input feature
-    - its own version of --recency
-    For prediction, it uses 'elapsed_days' for input to the forgetting curve.
-
-    This model with the default batch size (16384) uses a lot of memory.
-    If memory becomes a concern, use '--processes 1'.
-    Alternatively, reduce the batch size but the results would no longer be reproducible.
-
-    Just like with the GRU models, this model was trained on 100 users of the same dataset that it is tested on.
-    The effect on the resulting metrics is minor, but future work should be done to remove this influence.
-    """
-
-    def __init__(self, state_dict=None, input_mean=None, input_std=None):
-        super().__init__()
-        self.register_buffer(
-            "input_mean", torch.tensor(0.0) if input_mean is None else input_mean
-        )
-        self.register_buffer(
-            "input_std", torch.tensor(1.0) if input_std is None else input_std
-        )
-        self.n_input = 6
-        self.n_hidden = 20
-        self.n_curves = 3
-
-        self.process = nn.Sequential(
-            nn.Linear(self.n_input, self.n_hidden),
-            nn.SiLU(),
-            nn.LayerNorm((self.n_hidden,), bias=False),
-            nn.Linear(self.n_hidden, self.n_hidden),
-            nn.SiLU(),
-            ResBlock(
-                nn.Sequential(
-                    nn.LayerNorm((self.n_hidden,), bias=False),
-                    RNNWrapper(
-                        nn.LSTM(
-                            input_size=self.n_hidden,
-                            hidden_size=self.n_hidden,
-                            num_layers=1,
-                        )
-                    ),
-                )
-            ),
-            ResBlock(
-                nn.Sequential(
-                    nn.LayerNorm((self.n_hidden,)),
-                    RNNWrapper(
-                        nn.LSTM(
-                            input_size=self.n_hidden,
-                            hidden_size=self.n_hidden,
-                            num_layers=1,
-                        )
-                    ),
-                )
-            ),
-            ResBlock(
-                nn.Sequential(
-                    nn.LayerNorm((self.n_hidden,), bias=False),
-                    nn.Linear(self.n_hidden, self.n_hidden),
-                    nn.SiLU(),
-                    nn.LayerNorm((self.n_hidden,), bias=False),
-                    nn.Linear(self.n_hidden, self.n_hidden),
-                    nn.SiLU(),
-                )
-            ),
-            nn.LayerNorm((self.n_hidden,), bias=False),
-            nn.Linear(self.n_hidden, self.n_hidden),
-            nn.SiLU(),
-        )
-
-        for name, param in self.named_parameters():
-            if "weight_ih" in name:  # Input-to-hidden weights
-                nn.init.orthogonal_(param.data)
-            elif "weight_hh" in name:  # Hidden-to-hidden weights
-                nn.init.orthogonal_(param.data)
-            elif "bias_ih" in name:  # Biases
-                start_index = len(param.data) // 4
-                end_index = len(param.data) // 2
-                param.data[start_index:end_index].fill_(1.0)
-
-        self.w_fc = nn.Linear(self.n_hidden, self.n_curves)
-        self.s_fc = nn.Linear(self.n_hidden, self.n_curves)
-        self.d_fc = nn.Linear(self.n_hidden, self.n_curves)
-
-        if state_dict is not None:
-            self.load_state_dict(state_dict)
-        else:
-            try:
-                self.load_state_dict(
-                    torch.load(
-                        f"./pretrain/{FILE_NAME}_pretrain.pth",
-                        weights_only=True,
-                        map_location=DEVICE,
-                    )
-                )
-            except FileNotFoundError:
-                pass
-
-    def set_normalization_params(self, mean_i, std_i):
-        self.register_buffer("input_mean", mean_i)
-        self.register_buffer("input_std", std_i)
-
-    def forward(self, x_lni, hx=None):
-        x_delay, x_duration, x_rating = x_lni.split(1, dim=-1)
-        x_delay = torch.log(1e-5 + x_delay)
-        x_duration = torch.log(torch.clamp(x_duration, min=100, max=60000))
-        x_main = torch.cat([x_delay, x_duration], dim=-1)
-
-        x_main = (x_main - self.input_mean) / self.input_std
-
-        x_rating = torch.maximum(x_rating, torch.ones_like(x_rating))
-        x_rating = torch.nn.functional.one_hot(
-            x_rating.squeeze(-1).long() - 1, num_classes=4
-        ).float()
-        x = torch.cat([x_main, x_rating], dim=-1)
-        x_lnh = self.process(x)
-
-        w_lnh = torch.nn.functional.softmax(self.w_fc(x_lnh), dim=-1)
-        s_lnh = torch.exp(torch.clamp(self.s_fc(x_lnh), min=-25, max=25))
-        d_lnh = torch.exp(torch.clamp(self.d_fc(x_lnh), min=-25, max=25))
-        return w_lnh, s_lnh, d_lnh
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_n: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        w_lnh, s_lnh, d_lnh = self.forward(sequences)
-        (_, n, h) = w_lnh.shape
-        delta_nh = delta_n.unsqueeze(-1).expand(n, h)
-        w_nh = w_lnh[seq_lens - 1, torch.arange(n, device=DEVICE)]
-        s_nh = s_lnh[seq_lens - 1, torch.arange(n, device=DEVICE)]
-        d_nh = d_lnh[seq_lens - 1, torch.arange(n, device=DEVICE)]
-        retentions = self.forgetting_curve(delta_nh, w_nh, s_nh, d_nh)
-        return {"retentions": retentions, "stabilities": s_nh}
-
-    def forgetting_curve(self, t_nh, w_nh, s_nh, d_nh):
-        return (1 - 1e-7) * (
-            torch.sum(w_nh * (1 + t_nh / (1e-7 + s_nh)) ** -d_nh, dim=-1)
-        )
-
-
-class GRU_P(nn.Module):
-    # 297 params with default settings
-    lr: float = 1e-2
-    wd: float = 1e-5
-    n_epoch: int = 16
-
-    def __init__(self, state_dict=None):
-        super().__init__()
-        self.n_input = 2
-        self.n_hidden = 8
-        self.n_out = 1
-        self.n_layers = 1
-        self.rnn = nn.GRU(
-            input_size=self.n_input,
-            hidden_size=self.n_hidden,
-            num_layers=self.n_layers,
-        )
-        nn.init.orthogonal_(self.rnn.weight_ih_l0)
-        nn.init.orthogonal_(self.rnn.weight_hh_l0)
-        self.rnn.bias_ih_l0.data.fill_(0)
-        self.rnn.bias_hh_l0.data.fill_(0)
-
-        self.fc = nn.Linear(self.n_hidden, self.n_out)
-
-        if state_dict is not None:
-            self.load_state_dict(state_dict)
-        else:
-            try:
-                self.load_state_dict(
-                    torch.load(
-                        f"./pretrain/{FILE_NAME}_pretrain.pth",
-                        weights_only=True,
-                        map_location=DEVICE,
-                    )
-                )
-            except FileNotFoundError:
-                pass
-
-    def forward(self, x, hx=None):
-        x, h = self.rnn(x, hx=hx)
-        output = torch.sigmoid(self.fc(x))
-        return output, h
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        return {
-            "retentions": outputs[
-                seq_lens - 1, torch.arange(real_batch_size, device=DEVICE), 0
-            ]
-        }
-
-
-class Transformer(nn.Module):
-    # 127 params with default settings
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, state_dict=None):
-        super().__init__()
-        self.n_input = 2
-        self.n_hidden = 2
-        self.n_out = 1
-        self.n_layers = 1
-        self.transformer = nn.Transformer(
-            d_model=self.n_input,
-            nhead=self.n_input,
-            num_encoder_layers=self.n_layers,
-            num_decoder_layers=self.n_layers,
-            dim_feedforward=self.n_hidden,
-        )
-        self.fc = nn.Linear(self.n_input, self.n_out)
-
-        if state_dict is not None:
-            self.load_state_dict(state_dict)
-        else:
-            try:
-                self.load_state_dict(
-                    torch.load(
-                        f"./pretrain/{FILE_NAME}_pretrain.pth",
-                        weights_only=True,
-                        map_location=DEVICE,
-                    )
-                )
-            except FileNotFoundError:
-                pass
-
-    def forward(self, src):
-        tgt = torch.zeros(1, src.shape[1], self.n_input).to(device=DEVICE)
-        output = self.transformer(src, tgt)
-        output = self.fc(output)
-        output = torch.exp(output).repeat(src.shape[0], 1, 1)
-        return output, None
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        stabilities = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-            0,
-        ]
-        retentions = self.forgetting_curve(delta_ts, stabilities)
-        return {"retentions": retentions, "stabilities": stabilities}
-
-    def forgetting_curve(self, t, s):
-        return (1 + FACTOR * t / s) ** DECAY
-
-
-class HLR(nn.Module):
-    # 3 params
-    init_w = [2.5819, -0.8674, 2.7245]
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super().__init__()
-        self.n_input = 2
-        self.n_out = 1
-        self.fc = nn.Linear(self.n_input, self.n_out)
-
-        self.fc.weight = nn.Parameter(torch.tensor([w[:2]], dtype=torch.float32))
-        self.fc.bias = nn.Parameter(torch.tensor([w[2]], dtype=torch.float32))
-
-    def forward(self, x):
-        dp = self.fc(x)
-        return 2**dp
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs = self.forward(sequences.transpose(0, 1))
-        stabilities = outputs.squeeze(1)
-        retentions = self.forgetting_curve(delta_ts, stabilities)
-        return {"retentions": retentions, "stabilities": stabilities}
-
-    def forgetting_curve(self, t, s):
-        return 0.5 ** (t / s)
-
-    def state_dict(self):
-        return (
-            self.fc.weight.data.view(-1).tolist() + self.fc.bias.data.view(-1).tolist()
-        )
-
-
-class ACT_RParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp(0.001, 1)
-            w[1] = w[1].clamp(0.001, 1)
-            w[2] = w[2].clamp(0.001, 1)
-            w[3] = w[3].clamp_max(-0.001)
-            w[4] = w[4].clamp(0.001, 1)
-            module.w.data = w
-
-
-class ACT_R(nn.Module):
-    # 5 params
-    a = 0.176786766570677  # decay intercept
-    c = 0.216967308403809  # decay scale
-    s = 0.254893976981164  # noise
-    tau = -0.704205679427144  # threshold
-    h = 0.025  # interference scalar
-    init_w = [a, c, s, tau, h]
-    clipper = ACT_RParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super().__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def forward(self, sp: Tensor):
-        """
-        :param inputs: shape[seq_len, batch_size, 1]
-        """
-        m = torch.zeros_like(sp, dtype=torch.float)
-        m[0] = -torch.inf
-        for i in range(1, len(sp)):
-            act = torch.log(
-                torch.sum(
-                    ((sp[i] - sp[0:i]) * 86400 * self.w[4]).clamp_min(1)
-                    ** (-(self.w[1] * torch.exp(m[0:i]) + self.w[0])),
-                    dim=0,
-                )
-            )
-            m[i] = act
-        return self.activation(m[1:])
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs = self.forward(sequences)
-        return {
-            "retentions": outputs[
-                seq_lens - 2, torch.arange(real_batch_size, device=DEVICE), 0
-            ]
-        }
-
-    def activation(self, m):
-        return 1 / (1 + torch.exp((self.w[3] - m) / self.w[2]))
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class DASH(nn.Module):
-    # 9 params
-    if SHORT_TERM:
-        init_w = [
-            -0.1766,
-            0.4483,
-            -0.3618,
-            0.5953,
-            -0.5104,
-            0.8609,
-            -0.3643,
-            0.6447,
-            1.2815,
-        ]
-    else:
-        init_w = (
-            [0.2024, 0.5967, 0.1255, 0.6039, -0.1485, 0.572, 0.0933, 0.4801, 0.787]
-            if "MCM" not in MODEL_NAME
-            else [0.2783, 0.8131, 0.4252, 1.0056, -0.1527, 0.6455, 0.1409, 0.669, 0.843]
-        )
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(DASH, self).__init__()
-        self.fc = nn.Linear(8, 1)
-        self.sigmoid = nn.Sigmoid()
-
-        self.fc.weight = nn.Parameter(torch.tensor([w[:8]], dtype=torch.float32))
-        self.fc.bias = nn.Parameter(torch.tensor([w[8]], dtype=torch.float32))
-
-    def forward(self, x):
-        x = torch.log(x + 1)
-        x = self.fc(x)
-        x = self.sigmoid(x)
-        return x
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs = self.forward(sequences.transpose(0, 1))
-        return {"retentions": outputs.squeeze(1)}
-
-    def state_dict(self):
-        return (
-            self.fc.weight.data.view(-1).tolist() + self.fc.bias.data.view(-1).tolist()
-        )
-
-
-class DASH_ACTRParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp_min(0.001)
-            w[1] = w[1].clamp_min(0.001)
-            module.w.data = w
-
-
-class DASH_ACTR(nn.Module):
-    # 5 params
-    init_w = [1.4164, 0.516, -0.0564, 1.9223, 1.0549]
-    clipper = DASH_ACTRParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w=init_w):
-        super(DASH_ACTR, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-        self.sigmoid = nn.Sigmoid()
-
-    def forward(self, inputs):
-        """
-        :param inputs: shape[seq_len, batch_size, 2], 2 means r and t
-        """
-        inputs[:, :, 1] = inputs[:, :, 1].clamp_min(0.1)
-        retentions = self.sigmoid(
-            self.w[0]
-            * torch.log(
-                1
-                + torch.sum(
-                    torch.where(
-                        inputs[:, :, 1] == 0.1, 0, inputs[:, :, 1] ** -self.w[1]
-                    )
-                    * torch.where(inputs[:, :, 0] == 0, self.w[2], self.w[3]),
-                    dim=0,
-                ).clamp_min(0)
-            )
-            + self.w[4]
-        )
-        return retentions
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs = self.forward(sequences)
-        return {"retentions": outputs}
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class NN_17ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "S0"):
-            w = module.S0.data
-            w[0] = w[0].clamp(S_MIN, INIT_S_MAX)
-            w[1] = w[1].clamp(S_MIN, INIT_S_MAX)
-            w[2] = w[2].clamp(S_MIN, INIT_S_MAX)
-            w[3] = w[3].clamp(S_MIN, INIT_S_MAX)
-            module.S0.data = w
-
-        if hasattr(module, "D0"):
-            w = module.D0.data
-            w[0] = w[0].clamp(0, 1)
-            w[1] = w[1].clamp(0, 1)
-            w[2] = w[2].clamp(0, 1)
-            w[3] = w[3].clamp(0, 1)
-            module.D0.data = w
-
-        if hasattr(module, "sinc_w"):
-            w = module.sinc_w.data
-            w[0] = w[0].clamp(-5, 5)
-            w[1] = w[1].clamp(0, 1)
-            w[2] = w[2].clamp(-5, 5)
-            module.sinc_w.data = w
-
-
-def exp_activ(input):
-    return torch.exp(-input).clamp(0.0001, 0.9999)
-
-
-class ExpActivation(nn.Module):
-    def __init__(self):
-        super().__init__()  # init the base class
-
-    def forward(self, input):
-        return exp_activ(input)
-
-
-class NN_17(nn.Module):
-    # 39 params
-    init_s = [1, 2.5, 4.5, 10]
-    init_d = [1, 0.72, 0.07, 0.05]
-    w = [1.26, 0.0, 0.67]
-    clipper = NN_17ParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, state_dict=None) -> None:
-        super(NN_17, self).__init__()
-        self.hidden_size = 1
-        self.S0 = nn.Parameter(
-            torch.tensor(
-                self.init_s,
-                dtype=torch.float32,
-            )
-        )
-        self.D0 = nn.Parameter(
-            torch.tensor(
-                self.init_d,
-                dtype=torch.float32,
-            )
-        )
-        self.sinc_w = nn.Parameter(
-            torch.tensor(
-                self.w,
-                dtype=torch.float32,
-            )
-        )
-        self.rw = nn.Sequential(
-            nn.Linear(3, self.hidden_size),
-            nn.Mish(),
-            nn.Linear(self.hidden_size, 1),
-            # nn.Sigmoid()
-            nn.Softplus(),  # make sure that the input for ExpActivation() is >=0
-            ExpActivation(),
-        )
-        self.next_d = nn.Sequential(
-            nn.Linear(3, self.hidden_size),
-            nn.Mish(),
-            nn.Linear(self.hidden_size, 1),
-            nn.Sigmoid(),
-        )
-        self.pls = nn.Sequential(
-            nn.Linear(2, self.hidden_size),
-            nn.Mish(),
-            nn.Linear(self.hidden_size, 1),
-            nn.Softplus(),
-        )
-        self.sinc = nn.Sequential(
-            nn.Linear(3, self.hidden_size),
-            nn.Mish(),
-            nn.Linear(self.hidden_size, 1),
-            nn.Softplus(),
-        )
-        self.best_sinc = nn.Sequential(
-            nn.Linear(2, self.hidden_size),
-            nn.Mish(),
-            nn.Linear(self.hidden_size, 1),
-            nn.Softplus(),
-        )
-
-        if state_dict is not None:
-            self.load_state_dict(state_dict)
-        else:
-            try:
-                self.load_state_dict(
-                    torch.load(
-                        f"./pretrain/{FILE_NAME}_pretrain.pth",
-                        weights_only=True,
-                        map_location=DEVICE,
-                    )
-                )
-            except FileNotFoundError:
-                pass
-
-    def forward(self, inputs):
-        state = torch.ones((inputs.shape[1], 2), device=DEVICE)
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        stabilities = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-            0,
-        ]
-        difficulties = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-            1,
-        ]
-        theoretical_r = self.forgetting_curve(delta_ts, stabilities)
-        retentions = self.rw(
-            torch.stack([difficulties, stabilities, theoretical_r], dim=1)
-        ).squeeze(1)
-        return {"retentions": retentions, "stabilities": stabilities}
-
-    def step(self, X, state):
-        """
-        :param input: shape[batch_size, 3]
-            input[:,0] is elapsed time
-            input[:,1] is rating
-            input[:,2] is lapses
-        :param state: shape[batch_size, 2]
-            state[:,0] is stability
-            state[:,1] is difficulty
-        :return state:
-        """
-        delta_t = X[:, 0].unsqueeze(1)
-        rating = X[:, 1].unsqueeze(1)
-        lapses = X[:, 2].unsqueeze(1)
-
-        if torch.equal(state, torch.ones_like(state)):
-            # first review
-            keys = torch.tensor([1, 2, 3, 4], device=DEVICE)
-            keys = keys.view(1, -1).expand(X[:, 1].long().size(0), -1)
-            index = (X[:, 1].long().unsqueeze(1) == keys).nonzero(as_tuple=True)
-            new_s = torch.zeros_like(state[:, 0])
-            new_s[index[0]] = self.S0[index[1]]
-            new_s = new_s.unsqueeze(1)
-            new_d = torch.zeros_like(state[:, 1])
-            new_d[index[0]] = self.D0[index[1]]
-            new_d = new_d.unsqueeze(1)
-        else:
-            last_s = state[:, 0].unsqueeze(1)
-            last_d = state[:, 1].unsqueeze(1)
-
-            # Theoretical R
-            rt = self.forgetting_curve(delta_t, last_s)
-            rt = rt.clamp(0.0001, 0.9999)
-
-            # Rw
-            rw_input = torch.concat([last_d, last_s, rt], dim=1)
-            rw = self.rw(rw_input)
-            rw = rw.clamp(0.0001, 0.9999)
-
-            # S that corresponds to Rw
-            sr = self.inverse_forgetting_curve(rw, delta_t)
-            sr = sr.clamp(S_MIN, S_MAX)
-
-            # Difficulty
-            next_d_input = torch.concat([last_d, rw, rating], dim=1)
-            new_d = self.next_d(next_d_input)
-
-            # Post-lapse stability
-            pls_input = torch.concat([rw, lapses], dim=1)
-            pls = self.pls(pls_input)
-            pls = pls.clamp(S_MIN, S_MAX)
-
-            # SInc
-            sinc_t = 1 + torch.exp(self.sinc_w[0]) * (5 * (1 - new_d) + 1) * torch.pow(
-                sr, -self.sinc_w[1]
-            ) * torch.exp(-rw * self.sinc_w[2])
-
-            sinc_input = torch.concat([new_d, sr, rw], dim=1)
-            sinc_nn = 1 + self.sinc(sinc_input)
-            best_sinc_input = torch.concat([sinc_t, sinc_nn], dim=1)
-            best_sinc = 1 + self.best_sinc(best_sinc_input)
-            best_sinc.clamp(1, 100)
-
-            new_s = torch.where(
-                rating > 1,
-                sr * best_sinc,
-                pls,
-            )
-
-        new_s = new_s.clamp(S_MIN, S_MAX)
-        next_state = torch.concat([new_s, new_d], dim=1)
-        return next_state
-
-    def forgetting_curve(self, t, s):
-        return 0.9 ** (t / s)
-
-    def inverse_forgetting_curve(self, r: Tensor, t: Tensor) -> Tensor:
-        log_09 = -0.10536051565782628
-        return log_09 / torch.log(r) * t
-
-
-class SM2ParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            w[0] = w[0].clamp(S_MIN, S_MAX)
-            w[1] = w[1].clamp(S_MIN, S_MAX)
-            w[2] = w[2].clamp(1.3, 10.0)
-            w[3] = w[3].clamp(0, None)
-            w[4] = w[4].clamp(5, None)
-            module.w.data = w
-
-
-class SM2(nn.Module):
-    # 6 params
-    init_w = [1, 6, 2.5, 0.02, 7, 0.18]
-    clipper = SM2ParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(SM2, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def forward(self, inputs):
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        state = torch.zeros((inputs.shape[1], 3))  # [ivl, ef, reps]
-        state[:, 1] = self.w[2]
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        stabilities = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-            0,
-        ]
-        retentions = self.forgetting_curve(delta_ts, stabilities)
-        return {"retentions": retentions, "stabilities": stabilities}
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 3], state[:,0] is ivl, state[:,1] is ef, state[:,2] is reps
-        :return state: shape[batch_size, 3], [new_ivl, new_ef, new_reps]
-        """
-        rating = X[:, 1]
-        ivl = state[:, 0]
-        ef = state[:, 1]
-        reps = state[:, 2]
-        success = rating > 1
-
-        new_reps = torch.where(success, reps + 1, torch.ones_like(reps))
-        new_ivl = torch.where(
-            new_reps == 1,
-            self.w[0] * torch.ones_like(ivl),
-            torch.where(
-                new_reps == 2,
-                self.w[1] * torch.ones_like(ivl),
-                ivl * ef,
-            ),
-        )
-        q = rating + 1  # 1-4 -> 2-5
-        # EF':=EF+(0.1-(5-q)*(0.08+(5-q)*0.02))
-        # -> EF - 0.02 * (q-7) ^ 2 + 0.18
-        new_ef = ef - self.w[3] * (q - self.w[4]) ** 2 + self.w[5]
-        new_ivl = new_ivl.clamp(S_MIN, S_MAX)
-        new_ef = new_ef.clamp(1.3, 10.0)
-        return torch.stack([new_ivl, new_ef, new_reps], dim=1)
-
-    def forgetting_curve(self, t, s):
-        return 0.9 ** (t / s)
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
-
-
-class AnkiParameterClipper:
-    def __init__(self, frequency: int = 1):
-        self.frequency = frequency
-
-    def __call__(self, module):
-        if hasattr(module, "w"):
-            w = module.w.data
-            # based on limits in Anki 24.11
-            w[0] = w[0].clamp(1, 9999)
-            w[1] = w[1].clamp(1, 9999)
-            w[2] = w[2].clamp(1.31, 5.0)
-            w[3] = w[3].clamp(1, 5)
-            w[4] = w[4].clamp(0.5, 1.3)
-            w[5] = w[5].clamp(0, 1)
-            w[6] = w[6].clamp(0.5, 2)
-            module.w.data = w
-
-
-class Anki(nn.Module):
-    # 7 params
-    init_w = [
-        1,  # graduating interval
-        4,  # easy interval
-        2.5,  # starting ease
-        1.3,  # easy bonus
-        1.2,  # hard interval
-        0,  # new interval
-        1,  # interval multiplier
-    ]
-    clipper = AnkiParameterClipper()
-    lr: float = 4e-2
-    wd: float = 1e-5
-    n_epoch: int = 5
-
-    def __init__(self, w: List[float] = init_w):
-        super(Anki, self).__init__()
-        self.w = nn.Parameter(torch.tensor(w, dtype=torch.float32))
-
-    def forgetting_curve(self, t, s):
-        return 0.9 ** (t / s)
-
-    def passing_early_review_intervals(self, rating, ease, ivl, days_late):
-        elapsed = ivl + days_late
-        return torch.where(
-            rating == 2,
-            torch.max(elapsed * self.w[4], ivl * self.w[4] / 2),
-            torch.where(
-                rating == 4,
-                torch.max(elapsed * ease, ivl),
-                torch.max(elapsed * ease, ivl) * (self.w[3] - (self.w[3] - 1) / 2),
-            ),
-        )
-
-    def passing_nonearly_review_intervals(self, rating, ease, ivl, days_late):
-        return torch.where(
-            rating == 2,
-            ivl * self.w[4],
-            torch.where(
-                rating == 4,
-                (ivl + days_late / 2) * ease,
-                (ivl + days_late) * ease * self.w[3],
-            ),
-        )
-
-    def step(self, X: Tensor, state: Tensor) -> Tensor:
-        """
-        :param X: shape[batch_size, 2], X[:,0] is elapsed time, X[:,1] is rating
-        :param state: shape[batch_size, 2], state[:,0] is interval, state[:,1] is ease
-        :return state:
-        """
-        rating = X[:, 1]
-        if torch.equal(state, torch.zeros_like(state)):
-            # first learn, init memory states
-            new_ivl = torch.where(rating < 4, self.w[0], self.w[1])
-            new_ease = torch.ones_like(new_ivl) * self.w[2]
-        else:
-            ivl = state[:, 0]
-            ease = state[:, 1]
-            delta_t = X[:, 0]
-            days_late = delta_t - ivl
-            new_ivl = torch.where(
-                rating == 1,
-                ivl * self.w[5],
-                torch.where(
-                    days_late < 0,
-                    self.passing_early_review_intervals(rating, ease, ivl, days_late),
-                    self.passing_nonearly_review_intervals(
-                        rating, ease, ivl, days_late
-                    ),
-                )
-                * self.w[6],
-            )
-            new_ease = torch.where(
-                rating == 1,
-                ease - 0.2,
-                torch.where(
-                    rating == 2,
-                    ease - 0.15,
-                    torch.where(rating == 4, ease + 0.15, ease),
-                ),
-            )
-        new_ease = new_ease.clamp(1.3, 5.5)
-        new_ivl = torch.max(nn.functional.leaky_relu(new_ivl - 1) + 1, new_ivl).clamp(
-            S_MIN, S_MAX
-        )
-        return torch.stack([new_ivl, new_ease], dim=1)
-
-    def forward(
-        self, inputs: Tensor, state: Optional[Tensor] = None
-    ) -> tuple[Tensor, Tensor]:
-        """
-        :param inputs: shape[seq_len, batch_size, 2]
-        """
-        if state is None:
-            state = torch.zeros((inputs.shape[1], 2))
-        outputs = []
-        for X in inputs:
-            state = self.step(X, state)
-            outputs.append(state)
-        return torch.stack(outputs), state
-
-    def iter(
-        self,
-        sequences: Tensor,
-        delta_ts: Tensor,
-        seq_lens: Tensor,
-        real_batch_size: int,
-    ) -> dict[str, Tensor]:
-        outputs, _ = self.forward(sequences)
-        intervals = outputs[
-            seq_lens - 1,
-            torch.arange(real_batch_size, device=DEVICE),
-            0,
-        ]
-        retentions = self.forgetting_curve(delta_ts, intervals)
-        return {"retentions": retentions, "intervals": intervals}
-
-    def state_dict(self):
-        return list(
-            map(
-                lambda x: round(float(x), 4),
-                dict(self.named_parameters())["w"].data,
-            )
-        )
 
 
 def sm2(r_history):
@@ -3296,7 +994,7 @@ def process(user_id):
                     x = np.linspace(0, 1, len(train_partition))
                     train_partition["weights"] = 0.25 + 0.75 * np.power(x, 3)
 
-                model = Model()
+                model = Model(model_config)
                 if DRY_RUN:
                     partition_weights[partition] = model.state_dict()
                     continue
@@ -3344,7 +1042,9 @@ def process(user_id):
         for partition in testset["partition"].unique():
             partition_testset = testset[testset["partition"] == partition].copy()
             weights = w.get(partition, None)
-            my_collection = Collection(Model(weights) if weights else Model())
+            my_collection = Collection(
+                Model(model_config, weights) if weights else Model(model_config)
+            )
             retentions, stabilities, difficulties = my_collection.batch_predict(
                 partition_testset
             )
